# docs/api/openapi_v0.1.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "pakfindata API",
        "description": "Backend API for PSX market data. Endpoints under /v1 follow the Phase 1 contract (auth required, pydantic models). Endpoints under /api are legacy Phase-0 routers retained during the migration.",
        "version": "0.1.0"
    },
    "paths": {
...
        "/v1/market/kse100": {
            "get": {
                "tags": [
                    "market"
                ],
                "summary": "Kse100 Hero",
                "description": "Denormalized hero \u2014 KSE-100 quote, 52w range, and breadth.",
                "operationId": "kse100_hero_v1_market_kse100_get",
                "parameters": [
                    {
                        "name": "as_of",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Override the latest snapshot date",
                            "title": "As Of"
                        },
                        "description": "Override the latest snapshot date"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/KSE100Hero"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/market/top-gainers": {
            "get": {
                "tags": [
                    "market"
                ],
                "summary": "Top Gainers",
                "operationId": "top_gainers_v1_market_top_gainers_get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "default": 10,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Mover"
                                    },
                                    "title": "Response Top Gainers V1 Market Top Gainers Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/market/top-losers": {
            "get": {
                "tags": [
                    "market"
                ],
                "summary": "Top Losers",
                "operationId": "top_losers_v1_market_top_losers_get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "default": 10,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Mover"
                                    },
                                    "title": "Response Top Losers V1 Market Top Losers Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/market/volume-leaders": {
            "get": {
                "tags": [
                    "market"
                ],
                "summary": "Volume Leaders",
                "operationId": "volume_leaders_v1_market_volume_leaders_get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "default": 10,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Mover"
                                    },
                                    "title": "Response Volume Leaders V1 Market Volume Leaders Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/market/52w-extremes": {
            "get": {
                "tags": [
                    "market"
                ],
                "summary": "Fifty Two Week Extremes",
                "description": "Bundle of symbols near 52w high and near 52w low (Q3).",
                "operationId": "fifty_two_week_extremes_v1_market_52w_extremes_get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 50,
                            "minimum": 1,
                            "default": 5,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FiftyTwoWeekExtremes"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/market/value-leaders": {
            "get": {
                "tags": [
                    "market"
                ],
                "summary": "Value Leaders",
                "description": "Top symbols by turnover (close \u00d7 volume) for the latest trading day.",
                "operationId": "value_leaders_v1_market_value_leaders_get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "default": 10,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ValueLeader"
                                    },
                                    "title": "Response Value Leaders V1 Market Value Leaders Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/market/change-distribution": {
            "get": {
                "tags": [
                    "market"
                ],
                "summary": "Change Distribution",
                "description": "One row per symbol with rounded daily change % \u2014 feeds the\nMarket Pulse change-distribution histogram.",
                "operationId": "change_distribution_v1_market_change_distribution_get",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Trading date (YYYY-MM-DD); latest if omitted",
                            "title": "Date"
                        },
                        "description": "Trading date (YYYY-MM-DD); latest if omitted"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChangeDistributionRow"
                                    },
                                    "title": "Response Change Distribution V1 Market Change Distribution Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/market/announcements": {
            "get": {
                "tags": [
                    "market"
                ],
                "summary": "Announcements",
                "description": "Recent corporate announcements (company_announcements with a\ncorporate_announcements fallback).",
                "operationId": "announcements_v1_market_announcements_get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 200,
                            "minimum": 1,
                            "default": 8,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Announcement"
                                    },
                                    "title": "Response Announcements V1 Market Announcements Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ChangeDistributionRow": {
                "properties": {
                    "chg_pct": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Chg Pct"
                    }
                },
                "type": "object",
                "title": "ChangeDistributionRow",
                "description": "Histogram-bucket row from get_change_distribution.\n\nOne symbol \u2192 one rounded change percent. Clients bin these for the\nbreadth histogram."
            },
...
            "FiftyTwoWeekExtremes": {
                "properties": {
                    "near_high": {
                        "items": {
                            "$ref": "#/components/schemas/FiftyTwoWeekRow"
                        },
                        "type": "array",
                        "title": "Near High"
                    },
                    "near_low": {
                        "items": {
                            "$ref": "#/components/schemas/FiftyTwoWeekRow"
                        },
                        "type": "array",
                        "title": "Near Low"
                    }
                },
                "type": "object",
                "required": [
                    "near_high",
                    "near_low"
                ],
                "title": "FiftyTwoWeekExtremes",
                "description": "Bundle near-high + near-low \u2014 one widget, one round-trip."
            },
            "FiftyTwoWeekRow": {
                "properties": {
                    "symbol": {
                        "type": "string",
                        "title": "Symbol"
                    },
                    "pos_pct": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Pos Pct"
                    }
                },
                "type": "object",
                "required": [
                    "symbol"
                ],
                "title": "FiftyTwoWeekRow",
                "description": "Row shape from get_52w_extremes: ``symbol`` + position percent.\n\n``pos_pct`` is 100 at the 52-week high, 0 at the 52-week low."
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "KSE100Hero": {
                "properties": {
                    "as_of": {
                        "type": "string",
                        "title": "As Of"
                    },
                    "value": {
                        "type": "number",
                        "title": "Value"
                    },
                    "change": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Change"
                    },
                    "change_pct": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Change Pct"
                    },
                    "ytd_change_pct": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ytd Change Pct"
                    },
                    "one_year_change_pct": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "One Year Change Pct"
                    },
                    "week_52_high": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Week 52 High"
                    },
                    "week_52_low": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Week 52 Low"
                    },
                    "advancers": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Advancers"
                    },
                    "decliners": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Decliners"
                    },
                    "unchanged": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Unchanged"
                    }
                },
                "type": "object",
                "required": [
                    "as_of",
                    "value"
                ],
                "title": "KSE100Hero",
                "description": "Dashboard's hero widget \u2014 KSE-100 quote + breadth + 52w range.\n\nFields drawn from ``psx_indices`` (via ``get_latest_kse100``) plus\nbreadth from ``eod_market_summary`` (via ``get_eod_breadth``)."
            },
...
            "Mover": {
                "properties": {
                    "symbol": {
                        "type": "string",
                        "title": "Symbol"
                    },
                    "close": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Close"
                    },
                    "prev_close": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Prev Close"
                    },
                    "change_pct": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Change Pct"
                    },
                    "volume": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Volume"
                    }
                },
                "type": "object",
                "required": [
                    "symbol"
                ],
                "title": "Mover",
                "description": "Row shape returned by get_top_movers / get_volume_leaders."
            },
...
            "ValueLeader": {
                "properties": {
                    "symbol": {
                        "type": "string",
                        "title": "Symbol"
                    },
                    "close": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Close"
                    },
                    "volume": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Volume"
                    },
                    "value": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Value"
                    },
                    "change_pct": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Change Pct"
                    }
                },
                "type": "object",
                "required": [
                    "symbol"
                ],
                "title": "ValueLeader",
                "description": "Row shape from get_value_leaders \u2014 adds turnover to the Mover shape."
            }
        }
    }
}